+++ server/routes/facilityPhone.js
@@ -15,7 +15,7 @@ module.exports = function(app) {
   .post('/facility/:facilityID', function* () {
     const user = this.passport.user;
     const facilityID = this.params.facilityID;
-    if (user.scope.facilityID.toString() === facilityID) {
+    if (user.scope.facilityID && user.scope.facilityID.toString() === facilityID) {
       const requestJson = this.request.body;
       const q = {};
       q.sql = 'INSERT INTO ?? SET ?';
@@ -58,7 +58,7 @@ module.exports = function(app) {
   .put('/facility/:facilityID/phone/:phoneID', function* () {
     const user = this.passport.user;
     const facilityID = this.params.facilityID;
-    if (user.scope.facilityID.toString() === facilityID) {
+    if (user.scope.facilityID && user.scope.facilityID.toString() === facilityID) {
       const requestJson = this.request.body;
       const phoneID = this.params.phoneID;
       const q = {};
@@ -74,7 +74,7 @@ module.exports = function(app) {
   .delete('/facility/:facilityID/phone/:phoneID', function* () {
     const user = this.passport.user;
     const facilityID = this.params.facilityID;
-    if (user.scope.facilityID.toString() === facilityID) {
+    if (user.scope.facilityID && user.scope.facilityID.toString() === facilityID) {
       const phoneID = this.params.phoneID;
       const q = {};
       q.sql = 'DELETE FROM ?? WHERE ?? = ?';
